Fix typing

+++ src/backend/web/handlers/apidocs.py
@@ -2,7 +2,7 @@ from backend.common.decorators import cached_public
 from backend.web.profiled_render import render_template
 
 
-@cached_public(timeout=60 * 60 * 24 * 7)
+@cached_public(timeout=int(60 * 60 * 24 * 7))
 def apidocs_trusted_v1() -> str:
     template_values = {
         "title": "Trusted APIv1",
@@ -11,7 +11,7 @@ def apidocs_trusted_v1() -> str:
     return render_template("apidocs_swagger.html", template_values)
 
 
-@cached_public(timeout=60 * 60 * 24 * 7)
+@cached_public(timeout=int(60 * 60 * 24 * 7))
 def apidocs_v3() -> str:
     template_values = {
         "title": "APIv3",
